requirements.txt
streamlit==1.37.0
torch==2.3.0
torchvision==0.18.0
pillow==10.3.0
pymupdf==1.24.2
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.2
